requirements.txt
pymysql
mysql-replication
pytest

